refactor: move authentication and config files to core/ directories
PHASE 1 - Move authentication files:
- authentication.py → core/auth/authentication.py
- oauth_client.py → core/auth/oauth_client.py
- Update imports in source files

PHASE 2 - Move config files:
- config.py → core/config/config_manager.py
- ensemble_config.py → core/config/ensemble_config.py
- roles.py → core/config/roles.py
- Create core/config/ module structure
- Update imports in source files

STATUS: Source imports updated, test imports need fixing
NEXT: Continue reorganization of remaining files
