+++ src/llm_orc/orchestration.py
@@ -5,7 +5,7 @@ from typing import Any
 
 from llm_orc.communication import ConversationManager, MessageProtocol
 from llm_orc.models import ModelInterface
-from llm_orc.roles import RoleDefinition
+from llm_orc.core.config.roles import RoleDefinition
 
 
 class Agent:
